requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.0.0
pycparser==2.21
PyJWT==2.7.0
pymongo==4.3.3
pytest==7.3.1
python-dotenv==1.0.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.3.4
zipp==3.15.0
